consumer/requirements.txt
kafka-python
hdfs
prometheus_client
requests